+++ src/domain/market.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use std::collections::VecDeque;
 
@@ -5,6 +6,6 @@ use crate::domain::trades::Trade;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MarketData {
-    pub last_price: f64,
+    pub last_price: Decimal,
     pub last_updated: chrono::DateTime<chrono::Utc>,
 }
